[libcalamaresui] Move PythonQt.calamares.utils to libcalamares
Introduce a new top-level pre-loaded module into the PythonQt
binding called libcalamares. This will have the same API as the
Boost::Python version of the same, so that Python Job modules
using the "old" syntax continue to run with this library.

This affects PythonQt scripts, because calls to calamares.utils
need to be renamed. Since no API stability was promised, and
nothing has been created with these bindings, the change has
pretty much no impact. Just fix the example script.